New authentication flow
When the client create a new user it provides a challenge and its
answer. A challenge is onstentially an encrypted text that only the
client can decrypt.
When the client need to authenticate it must request the challenge and
send back the answer. If it matches with the one initially provided then
the client is allowed.